Refactor to convert old core actions to aliases and the new actions to hold the code.
