+++ fastlane/lib/fastlane/actions/check_app_store_metadata.rb
@@ -1,18 +1,51 @@
 module Fastlane
   module Actions
-    require 'fastlane/actions/precheck'
-    class CheckAppStoreMetadataAction < PrecheckAction
+    module SharedValues
+    end
+
+    class CheckAppStoreMetadataAction < Action
       def self.run(config)
-        UI.message "Checking App Store metadata using precheck"
-        super(config)
+        require 'precheck'
+        Precheck.config = config
+        return Precheck::Runner.new.run
       end
 
-      #####################################################
-      # @!group Documentation
-      #####################################################
-
       def self.description
-        "Check your app's metadata to avoid being rejected (via precheck)"
+        "Check your app's metadata before you submit your app to review (via `precheck`)"
+      end
+
+      def self.details
+        "More information: https://fastlane.tools/precheck"
+      end
+
+      def self.available_options
+        require 'precheck/options'
+        Precheck::Options.available_options
+      end
+
+      def self.return_value
+        return "true if precheck passes, else, false"
+      end
+
+      def self.authors
+        ["taquitos"]
+      end
+
+      def self.is_supported?(platform)
+        platform == :ios
+      end
+
+      def self.example_code
+        [
+          'check_app_store_metadata(
+            negative_apple_sentiment(level: :skip), # Set to skip to not run the `negative_apple_sentiment` rule
+            curse_words(level: :warn) # Set to warn to only warn on curse word check failures
+          )'
+        ]
+      end
+
+      def self.category
+        :misc
       end
     end
   end
