+++ fastlane/lib/fastlane/actions/create_itunes_connect_app.rb
@@ -1,18 +1,74 @@
 module Fastlane
   module Actions
-    require 'fastlane/actions/produce'
-    class CreateItunesConnectAppAction < ProduceAction
-      def self.run(config)
-        UI.message "Creating a new app using produce"
-        super(config)
+    module SharedValues
+      PRODUCE_APPLE_ID = :PRODUCE_APPLE_ID
+    end
+
+    class CreateItunesConnectAppAction < Action
+      def self.run(params)
+        require 'produce'
+
+        return if Helper.test?
+
+        Produce.config = params # we alread have the finished config
+
+        Dir.chdir(FastlaneCore::FastlaneFolder.path || Dir.pwd) do
+          # This should be executed in the fastlane folder
+          apple_id = Produce::Manager.start_producing.to_s
+
+          Actions.lane_context[SharedValues::PRODUCE_APPLE_ID] = apple_id
+          ENV['PRODUCE_APPLE_ID'] = apple_id
+        end
       end
 
-      #####################################################
-      # @!group Documentation
-      #####################################################
-
       def self.description
-        "Creates a new app on iTunes Connect and the Apple Developer Portal (via produce)"
+        "Creates the given application on iTC and the Dev Portal (via `produce`)"
+      end
+
+      def details
+        [
+          'Create new apps on iTunes Connect and Apple Developer Portal via `produce`.',
+          'If the app already exists, `create_itunes_connect_app` will not do anything.',
+          'For more information about produce, visit its GitHub page:',
+          'https://github.com/fastlane/fastlane/tree/master/produce'
+        ].join(' ')
+      end
+
+      def self.available_options
+        require 'produce'
+        Produce::Options.available_options
+      end
+
+      def self.output
+        [
+          ['PRODUCE_APPLE_ID', 'The Apple ID of the newly created app. You probably need it for `deliver`']
+        ]
+      end
+
+      def self.author
+        "KrauseFx"
+      end
+
+      def self.is_supported?(platform)
+        platform == :ios
+      end
+
+      def self.example_code
+        [
+          'create_itunes_connect_app(
+            username: "felix@krausefx.com",
+            app_identifier: "com.krausefx.app",
+            app_name: "MyApp",
+            language: "English",
+            app_version: "1.0",
+            sku: "123",
+            team_name: "SunApps GmbH" # Only necessary when in multiple teams.
+          )'
+        ]
+      end
+
+      def self.category
+        :misc
       end
     end
   end
