+++ fastlane/lib/fastlane/actions/frame_screenshots.rb
@@ -1,18 +1,60 @@
 module Fastlane
   module Actions
-    require 'fastlane/actions/frameit'
-    class FrameScreenshotsAction < FrameitAction
+    class FrameScreenshotsAction < Action
       def self.run(config)
-        UI.message "Framing screenshots using frameit"
-        super(config)
+        return if Helper.test?
+
+        require 'frameit'
+
+        UI.message("Framing screenshots at path #{config[:path]} (via frameit)")
+
+        Dir.chdir(config[:path]) do
+          Frameit.config = config
+          Frameit::Runner.new.run('.')
+        end
       end
 
-      #####################################################
-      # @!group Documentation
-      #####################################################
-
       def self.description
-        "Adds a black frame around all screenshots (via frameit)"
+        "Adds device frames around all screenshots (via `frameit`)"
+      end
+
+      def self.details
+        [
+          "Uses [frameit](https://github.com/fastlane/fastlane/tree/master/frameit) to prepare perfect screenshots for the App Store, your website, QA",
+          "or emails. You can add background and titles to the framed screenshots as well."
+        ].join("\n")
+      end
+
+      def self.available_options
+        require "frameit"
+        require "frameit/options"
+        FastlaneCore::CommanderGenerator.new.generate(Frameit::Options.available_options) + [
+          FastlaneCore::ConfigItem.new(key: :path,
+                                       env_name: "FRAMEIT_SCREENSHOTS_PATH",
+                                       description: "The path to the directory containing the screenshots",
+                                        default_value: Actions.lane_context[SharedValues::SNAPSHOT_SCREENSHOTS_PATH] || FastlaneCore::FastlaneFolder.path)
+        ]
+      end
+
+      def self.author
+        "KrauseFx"
+      end
+
+      def self.example_code
+        [
+          'frame_screenshots',
+          'frame_screenshots(silver: true)',
+          'frame_screenshots(path: "/screenshots")',
+          'frame_screenshots(rose_gold: true)'
+        ]
+      end
+
+      def self.category
+        :screenshots
+      end
+
+      def self.is_supported?(platform)
+        [:ios, :mac].include? platform
       end
     end
   end
