+++ fastlane/lib/fastlane/actions/screengrab.rb
@@ -1,55 +1,13 @@
 module Fastlane
   module Actions
-    module SharedValues
-      SCREENGRAB_OUTPUT_DIRECTORY = :SCREENGRAB_OUTPUT_DIRECTORY
-    end
-
-    class ScreengrabAction < Action
-      def self.run(params)
-        require 'screengrab'
-
-        Screengrab.config = params
-        Screengrab.android_environment = Screengrab::AndroidEnvironment.new(params[:android_home],
-                                                                            params[:build_tools_version])
-        Screengrab::DependencyChecker.check(Screengrab.android_environment)
-        Screengrab::Runner.new.run
-
-        Actions.lane_context[SharedValues::SCREENGRAB_OUTPUT_DIRECTORY] = File.expand_path(params[:output_directory])
-
-        true
-      end
+    require 'fastlane/actions/capture_android_screenshots'
+    class ScreengrabAction < CaptureAndroidScreenshotsAction
+      #####################################################
+      # @!group Documentation
+      #####################################################
 
       def self.description
-        'Automated localized screenshots of your Android app on every device'
-      end
-
-      def self.available_options
-        require 'screengrab'
-        Screengrab::Options.available_options
-      end
-
-      def self.author
-        ['asfalcone', 'i2amsam', 'mfurtak']
-      end
-
-      def self.is_supported?(platform)
-        platform == :android
-      end
-
-      def self.example_code
-        [
-          'screengrab',
-          'screengrab(
-            locales: ["en-US", "fr-FR", "ja-JP"],
-            clear_previous_screenshots: true,
-            app_apk_path: "build/outputs/apk/example-debug.apk",
-            tests_apk_path: "build/outputs/apk/example-debug-androidTest-unaligned.apk"
-          )'
-        ]
-      end
-
-      def self.category
-        :screenshots
+        "Alias for the `capture_android_screenshots` action"
       end
     end
   end
